+++ NEWS.md
@@ -1,3 +1,12 @@
+# causaloptim 1.0.0
+
+Major release with lots of changes: 
+
+- Balke's legacy C++ code is no longer supported. Vertex enumeration is now always done with rcdd
+- New interface to specify a causal model without an effect. Also derivation of observable constraints
+- Various bug fixes and improvements
+- New vignette to showcase new features
+
 # causaloptim 0.10.0
 
 - experimental functionality for automated linearity checking